tests/wpt/tests/paint-timing/first-contentful-canvas.html

<!DOCTYPE html>
<head>
<title>Performance Paint Timing Test: FCP due to canvas</title>
</head>
<body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/utils.js"></script>
<canvas id="canvas" width="200" height="200" ></canvas>

<script>
setup({"hide_test_state": true});
async_test(function (t) {
    assert_implements(window.PerformancePaintTiming, "Paint Timing isn't supported.");
    const canvas = document.getElementById("canvas");
    const context = canvas.getContext("2d");
    context.beginPath();
    context.moveTo(0,0);
    context.lineTo(300,150);
    context.stroke();
    assertPaintTimingEntries(t, 20);
}, "First contentful paint fires due to canvas render.");
</script>
</body>
</html>
